Add exploding death to enemies

+++ Assets/Player/Scripts/PlayerInfo.cs
@@ -5,8 +5,8 @@ using UnityEngine;
   
 public class PlayerInfo
 {
-    private int maxHealth = 10000;
-    private int currentHealth = 0;
+    private float maxHealth = 10000;
+    private float currentHealth = 0;
 
     private int currentAtonement = 0;
     private int atonementLvl = 0;
@@ -14,8 +14,8 @@ public class PlayerInfo
 
    
    
-    public int MaxHealth { get => maxHealth; set => maxHealth = value; }
-    public int CurrentHealth { get => currentHealth; set => currentHealth = value; }
+    public float MaxHealth { get => maxHealth; set => maxHealth = value; }
+    public float CurrentHealth { get => currentHealth; set => currentHealth = value; }
     public int CurrentAtonement { get => currentAtonement; set => currentAtonement = value; }
     public int AtonementLvl { get => atonementLvl; set => atonementLvl = value; }
     public int AtonementToLevel { get => atonementToLevel; set => atonementToLevel = value; }
